Widespread Improvements.
- Build system made more robust, now provides a macro to indicate
NVDECODE/Video Codec SDK version.
- Build system now uses same Python installation as setuptools.
- Bugfixes for Python object reference leaks and unprotected uses:
  - Method waitToken() added to extract Python objects pushed into but
not out of an iterator core
  - Constructor of iterator core now asks for multibuffer object in
addition to its base pointer, so that it can hold a reference to it for
the lifetime of the bmultibuffer. This prevents it from being
deallocated from underneath the iterator core, as can happen if
destructors delete the buffer first and the core second.
- Added a benchmark script.