Solve critical race condition leading to segfaults.
The destruction of an NVDECODE_CTX with nvdecodeRelease() already
guaranteed that all threads had "exited" and released the mutex. The
problem is that it did not guarantee that the threads had actually
completed executing and been joined. Thus, if nvdecodeRelease() was
followed in short order by dlclose() of the plugin, the thread could see
its own code be unmapped from underneath itself, and this rug-pulling
resulted in SIGSEGV as the thread was being woken up a final time.

We solve this by making the threads joinable, introducing a new CTX
state, and joining the helpers before exiting.