Solve memory leak problems by supporting cyclic garbage collection.
This implies the following new slots:

- tp_alloc
- tp_free
- tp_traverse
- tp_clear
- tp_finalize

The iterator core may potentially become part of cyclic trash keeping
the multibuffer alive, but the garbage collector does not immediately
free cyclic isolates. To solve this, at the first and last calls to
next() on the iterator, we:

  1. Destroy the multibuffer and iterator core attributes
  2. Cause the GC to run and reap cyclic isolates
  3. Empty the PyTorch CUDA cache.

The combined effect is to release any CUDA memory that can be released,
and return it to the PyTorch memory pool.