Fix dlopen() of libbenzina.so on Glibc >= 2.30
Tests were broken because Glibc 2.30 added a new check to prevent
dynamic loading of PIE executables. dlopen() now checks for a flag
DF_1_PIE in the .dynamic section/DYNAMIC segment's DT_FLAGS_1 tag and
refuses to dlopen() it at all.

The linkers set this flag when building -pie and there is no way to
disable it, so we manually patch it out with Python code. This requires
direct, ugly manipulation of ELF files but, happily, requires only
architecture-independent logic.